Ensure JUL OFF level is respected in liquibase logger
Motivation
----------

An issue was found where the JUL OFF level was being logged at the error level. This is due to how the log level is determined using integer value checks. The ERROR level is the fallback when it should not be.

Modifications
-------------

A reproducer test was written for the issue and the ERROR threshold was checked to fix the issue. The OFF level is now the fallback which does nothing.

Result
------

The result is the JUL OFF level is now respected.